Lint

+++ pkg/kube/krt/join_test.go
@@ -315,7 +315,7 @@ func TestJoinWithMergeCollection(t *testing.T) {
 			for p := range ports {
 				finalServiceInfo.Service.Ports = append(finalServiceInfo.Service.Ports, &workloadapi.Port{
 					ServicePort: p.port,
-					TargetPort:  uint32(p.targetPort),
+					TargetPort:  p.targetPort,
 				})
 			}
